refactor(view:matter): moved type files
核となる要素のActor, World, Controllerと
それに付随する要素を別々のディレクトリに分けた
