+++ src/view/matter/types/core/controller.ts
@@ -1,7 +1,7 @@
 /* eslint-disable class-methods-use-this */
 import P5Types from 'p5';
-import { World } from 'src/view/matter/types/world';
-import { ActorParameter } from 'src/view/matter/types/actorParameter';
+import { World } from 'src/view/matter/types/core/world';
+import { ActorParameter } from 'src/view/matter/types/util/actorParameter';
 
 /**
  * {@link Actor}の振る舞いを変化させる
